chore: rename capacity_value.py to capacity.py
